#1694 - Improve annotation detail sidebar architecture and performance
- Pull out code for the feature panel into a separate component
