+++ webanno-ui-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/annotation/detail/AnnotationInfoPanel.html
@@ -19,26 +19,31 @@
 <html xmlns:wicket="http://wicket.apache.org">
 <body>
   <wicket:panel>
-    <div class="form-group form-row d-flex" wicket:enclosure="selectedAnnotationType">
-      <div class="flex-fill form-control scrolling" style="height: unset; max-height: 4.5rem; word-wrap: break-word;" readonly>
-        <div class="small" wicket:id="selectedAnnotationType"/>
-      </div>
+    <div wicket:id="noAnnotationWarning" class="flex-content flex-h-container no-data-notice">
+      <div>No annotation selected</div>
     </div>
-    <div class="feature-editors-sidebar">
-      <div class="form-group form-row" wicket:enclosure="selectedAnnotationLayer">
-        <label class="feature-editor-label col-form-label">Layer</label>
-        <div class="feature-editor-value">
-          <span class="form-control" readonly><wicket:container wicket:id="selectedAnnotationLayer"/></span>
+    <div wicket:id="annotationInfo">
+      <div class="form-group form-row d-flex" wicket:enclosure="selectedAnnotationType">
+        <div class="flex-fill form-control scrolling" style="height: unset; max-height: 4.5rem; word-wrap: break-word;" readonly>
+          <div class="small" wicket:id="selectedAnnotationType"/>
         </div>
       </div>
-      <div class="form-group form-row">
-        <label class="feature-editor-label col-form-label">
-          <wicket:message key="text"/> 
-          <a wicket:id="jumpToAnnotation" class="float-right"><i class="fas fa-crosshairs"></i></a>
-        </label>
-        
-        <div class="feature-editor-value form-control scrolling" style="height: unset; max-height: 4.5rem; word-wrap: break-word;" readonly>
-          <wicket:container wicket:id="selectedText"/>
+      <div class="feature-editors-sidebar">
+        <div class="form-group form-row" wicket:enclosure="selectedAnnotationLayer">
+          <label class="feature-editor-label col-form-label">Layer</label>
+          <div class="feature-editor-value">
+            <span class="form-control" readonly><wicket:container wicket:id="selectedAnnotationLayer"/></span>
+          </div>
+        </div>
+        <div class="form-group form-row">
+          <label class="feature-editor-label col-form-label">
+            <wicket:message key="text"/> 
+            <a wicket:id="jumpToAnnotation" class="float-right"><i class="fas fa-crosshairs"></i></a>
+          </label>
+          
+          <div class="feature-editor-value form-control scrolling" style="height: unset; max-height: 4.5rem; word-wrap: break-word;" readonly>
+            <wicket:container wicket:id="selectedText"/>
+          </div>
         </div>
       </div>
     </div>
