+++ webanno-ui-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/annotation/detail/LayerSelectionPanel.java
@@ -209,7 +209,7 @@ public class LayerSelectionPanel
             }
         }));
         checkbox.add(new LambdaAjaxFormComponentUpdatingBehavior("change", _target -> 
-                owner.focusForwardAnnotationComponent(_target, true)));
+                owner.getFeatureEditorContainer().focusForwardAnnotationComponent(_target, true)));
         
         return checkbox;
     }
